Fix hasRunning
The last version could get maxRunDuration of zero when we have
waiting or failed trials without any complete or pruned trials.
It leads to false in the judge of hasRunning.
However, in this case, we should be able to say that we have
running trials.
I fixed this issue in this commit.